WPF Fix `lParam.ToInt32()` overflow when running in 64bit mode https://github.com/cefsharp/CefSharp/issues/509#issuecomment-70944315

+++ CefSharp.Wpf/ChromiumWebBrowser.cs
@@ -625,7 +625,7 @@ namespace CefSharp.Wpf
                         return IntPtr.Zero;
                     }
 
-                    handled = managedCefBrowserAdapter.SendKeyEvent(message, wParam.ToInt32(), lParam.ToInt32());
+                    handled = managedCefBrowserAdapter.SendKeyEvent(message, wParam.ToInt32(), lParam.ToInt64());
                     
                     break;
                 }
@@ -898,7 +898,7 @@ namespace CefSharp.Wpf
                 var message = (int)(e.IsDown ? WM.KEYDOWN : WM.KEYUP);
                 var virtualKey = KeyInterop.VirtualKeyFromKey(e.Key);
 
-                e.Handled = managedCefBrowserAdapter.SendKeyEvent(message, virtualKey, (int)modifiers);
+                e.Handled = managedCefBrowserAdapter.SendKeyEvent(message, virtualKey, (Int64)modifiers);
             }
         }
